PromMetrics: move to metrics.{cc/hh}
