+++ src/hydra-queue-runner/state.hh
@@ -7,11 +7,6 @@
 #include <memory>
 #include <queue>
 
-#include <prometheus/counter.h>
-#include <prometheus/gauge.h>
-#include <prometheus/histogram.h>
-#include <prometheus/registry.h>
-
 #include "db.hh"
 
 #include "parsed-derivations.hh"
@@ -21,6 +16,7 @@
 #include "store-api.hh"
 #include "sync.hh"
 #include "nar-extractor.hh"
+#include "metrics.hh"
 
 
 typedef unsigned int BuildID;
@@ -230,22 +226,6 @@ void getDependents(Step::ptr step, std::set<Build::ptr> & builds, std::set<Step:
 /* Call ‘visitor’ for a step and all its dependencies. */
 void visitDependencies(std::function<void(Step::ptr)> visitor, Step::ptr step);
 
-struct PromTimer
-{
-    std::chrono::time_point<std::chrono::high_resolution_clock> created;
-
-    PromTimer()
-        : created(std::chrono::high_resolution_clock::now())
-    {
-
-    }
-
-public:
-    void finish(prometheus::Histogram& histogram)
-    {
-        histogram.Observe(std::chrono::duration<double>(std::chrono::high_resolution_clock::now() - created).count());
-    }
-};
 
 struct Machine
 {
@@ -456,24 +436,6 @@ private:
 
     std::string metricsAddr;
 
-    struct PromMetrics
-    {
-        std::shared_ptr<prometheus::Registry> registry;
-
-        prometheus::Counter& queue_checks_started;
-        prometheus::Histogram& queue_build_fetch_time;
-        prometheus::Family<prometheus::Histogram>& queue_build_load_family;
-        prometheus::Histogram& queue_build_load_premature_gc;
-        prometheus::Histogram& queue_build_load_cached_failure;
-        prometheus::Histogram& queue_build_load_cached_success;
-        prometheus::Histogram& queue_build_load_added;
-        prometheus::Counter& queue_steps_created;
-        prometheus::Counter& queue_checks_early_exits;
-        prometheus::Counter& queue_checks_finished;
-        prometheus::Gauge& queue_max_id;
-
-        PromMetrics();
-    };
     PromMetrics prom;
 
 public:
